pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "finsight-agents"
version = "0.1.0"
description = "Multi-agent financial insight system demonstrating LangGraph, MCP, evaluations, OTel, and drift detection"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.3.2",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27.0",
    "prometheus-client>=0.21.0",
    "mlflow>=2.16.0",
    "scipy>=1.14.0",
    "pyyaml>=6.0.2",
    "jinja2>=3.1.4",
    "streamlit>=1.39.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
    "honcho>=2.0.0",
    "click>=8.1.7",
    "anyio>=4.4.0",
]

[project.scripts]
finsight = "app.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]
omit = ["app/ui/*"]
